added manual mocks for clerk
