+++ tests/pages/api/event/[id]/update.test.ts
@@ -1,8 +1,8 @@
 import type { PageConfig } from "next";
 import { testApiHandler } from "next-test-api-route-handler";
-import { beforeAll, describe, expect, it } from "vitest";
+import { beforeAll, describe, expect, it, vi } from "vitest";
 
-import { cleanup, seedTestData, setupClerkMocks } from "@/../tests/utils";
+import { cleanup, seedTestData } from "@/../tests/utils";
 import prisma from "@/lib/prisma";
 import endpoint from "@/pages/api/event/[id]";
 
@@ -13,7 +13,7 @@ describe("PATCH /api/event/:id", () => {
   beforeAll(async () => {
     await cleanup();
     await seedTestData();
-    setupClerkMocks();
+    vi.mock("@clerk/nextjs/server");
   });
 
   it("should return 400 status code on invalid fields", async () => {
