+++ tests/pages/api/event/create.test.ts
@@ -1,11 +1,11 @@
 import type { PageConfig } from "next";
 import { testApiHandler } from "next-test-api-route-handler";
-import { beforeAll, describe, expect, it } from "vitest";
+import { beforeAll, describe, expect, it, vi } from "vitest";
 
 import prisma from "@/lib/prisma";
 import endpoint from "@/pages/api/event";
 
-import { cleanup, seedTestData, setupClerkMocks } from "../../../utils";
+import { cleanup, seedTestData } from "../../../utils";
 
 // Respect the Next.js config object if it's exported
 const handler: typeof endpoint & { config?: PageConfig } = endpoint;
@@ -14,7 +14,7 @@ describe("POST /api/event", () => {
   beforeAll(async () => {
     await cleanup();
     await seedTestData();
-    setupClerkMocks();
+    vi.mock("@clerk/nextjs/server");
   });
 
   it("should return 400 status code on invalid fields", async () => {
